+++ src/lsp/cobol_lsp/lsp_data_info_printer.ml
@@ -68,36 +68,47 @@ let pp_usage: usage Pretty.printer =
   let pp_usage_with_picture ppf name (picture: Cobol_data.Picture.t) =
     Fmt.pf ppf "%a\n\n%a%a"
       (pp_cobol_block (fun ppf _ ->
-           Fmt.pf ppf "PIC %a USAGE %s"
+           Fmt.pf ppf ("PIC %a USAGE " ^^ name)
              Cobol_data.Picture.pp_picture_symbols picture.pic
-             name)) ()
+         )) ()
       Cobol_data.Picture.pp_category picture.category
       pp_example_of picture
-  and pp_usage_with_sign ppf name signed =
-    pp_cobol_block Fmt.(any "USAGE " ++ any name ++
-                        any (if signed then " SIGNED" else " UNSIGNED"))
-      ppf ()
   and pp_width_tag ppf tag =
     Fmt.int ppf @@
     match tag with `W16 -> 16 | `W32 -> 32 | `W34 -> 34 | `W64 -> 64 | `W128 -> 128
   in
+  let pp_usage_with_optional_picture ppf name signed picture =
+    match picture with
+    | Some picture ->
+        pp_usage_with_picture ppf name picture
+    | None ->
+        pp_cobol_block Fmt.(any "USAGE " ++ any name ++
+                            any (if signed then " SIGNED" else " UNSIGNED"))
+          ppf ()
+  in
   fun ppf -> function
-    | Binary picture ->
-        pp_usage_with_picture ppf "BINARY" picture
-    | Binary_C_long { signed } ->
-        pp_usage_with_sign ppf "BINARY-C-LONG" signed
-    | Binary_char { signed } ->
-        pp_usage_with_sign ppf "BINARY-CHAR" signed
-    | Binary_double { signed; _ } ->
-        pp_usage_with_sign ppf "BINARY-DOUBLE" signed
-    | Binary_long { signed; _ } ->
-        pp_usage_with_sign ppf "BINARY-LONG" signed
-    | Binary_short { signed; _ } ->
-        pp_usage_with_sign ppf "BINARY-SHORT" signed
+    | Alphanumeric { picture; _ }
+    | Display_numeric { picture; _ } ->
+        pp_usage_with_picture ppf "DISPLAY" picture
+    | Binary { picture; byte_size = Byte_size; signed;
+               truncation = Truncate_to_native_size; _ } ->
+        pp_usage_with_optional_picture ppf "BINARY-CHAR" signed picture
+    | Binary { picture; byte_size = Short_size; signed;
+               truncation = Truncate_to_native_size; _ } ->
+        pp_usage_with_optional_picture ppf "BINARY-SHORT" signed picture
+    | Binary { picture; byte_size = Long_size; signed;
+               truncation = Truncate_to_native_size; _ } ->
+        pp_usage_with_optional_picture ppf "BINARY-LONG" signed picture
+    | Binary { picture; byte_size = Double_size; signed;
+               truncation = Truncate_to_native_size; _ } ->
+        pp_usage_with_optional_picture ppf "BINARY-DOUBLE" signed picture
+    | Binary { picture; byte_size = C_long_size; signed;
+               truncation = Truncate_to_native_size; _ } ->
+        pp_usage_with_optional_picture ppf "BINARY-C-LONG" signed picture
+    | Binary { picture; signed; _ } ->
+        pp_usage_with_optional_picture ppf "BINARY" signed picture
     | Bit picture ->
         pp_usage_with_picture ppf "BIT" picture
-    | Display picture ->
-        pp_usage_with_picture ppf "DISPLAY" picture
     | Float_binary { width; endian = _ } ->
         Pretty.record [
           Fmt.(styled `Yellow @@ any "float-binary");