+++ src/lsp/cobol_typeck/typeck_data_diagnostics.ml
@@ -72,7 +72,7 @@ type error =
       }
   | Missing_picture_clause_for_item_with_usage of
       {
-        usage_clause: Cobol_ptree.usage_clause;
+        usage: Cobol_ptree.usage_clause;
         item_loc: srcloc;
       }
   | Unexpected_picture_clause of
@@ -119,14 +119,15 @@ type error =
   | Incompatible_picture of
       {
         picture: Cobol_data.Picture.t with_loc;
-        usage_clause: Cobol_ptree.usage_clause;
-        expected: [`Numeric_category | `Boolean_class | `Nonalpha_class];
+        usage: Cobol_ptree.usage_clause;
+        expected: [`Numeric_category | `Numeric_or_alphanum_category
+                  | `Boolean_class | `Nonalpha_class];
       }
   | Invalid_picture_feature of
       {
         picture: Cobol_data.Picture.t with_loc;
         feature: picture_feature;
-        usage_clause: Cobol_ptree.usage_clause;
+        usage: Cobol_ptree.usage_clause;
       }
   | Item_not_found of
       {
@@ -165,6 +166,11 @@ and warning =
         clause_loc: srcloc;
         clause_name: string;
       }
+  | Ignored_picture_clause of
+      {
+        picture: Cobol_data.Types.picture with_loc;
+        usage: Cobol_ptree.usage_clause;
+      }
   | Mismatching_usage_in_group of                 (* not enforced by GnuCOBOL *)
       {
         item_name: Cobol_ptree.data_name with_loc option;
@@ -179,7 +185,7 @@ and warning =
       }
   | Unsupported_usage of
       {
-        usage_clause: Cobol_ptree.usage_clause with_loc;
+        usage: Cobol_ptree.usage_clause with_loc;
       }
   (* | Extraneous_clause of *)
   (*     { *)
@@ -225,9 +231,10 @@ let error_loc = function
 let warning_loc = function
   | Duplicate_clause { second_loc = loc; _ }
   | Ignored_clause_in_redefinition { clause_loc = loc; _ }
+  | Ignored_picture_clause { picture = { loc; _ }; _ }
   | Mismatching_usage_in_group { item_usage = { loc; _ }; _ }
   | Redefinition_of_table_item { redef_loc = loc; _ }
-  | Unsupported_usage { usage_clause = { loc; _ } } ->
+  | Unsupported_usage { usage = { loc; _ } } ->
       Some loc
 
 let pp_data_name'_opt
@@ -247,23 +254,24 @@ let pp_picture_feature_kind ppf = function
   | Scaling _ -> Fmt.pf ppf "scaling"
 
 let pp_error ppf = function
-  | Incompatible_picture { picture; usage_clause; expected } ->
+  | Incompatible_picture { picture; usage; expected } ->
       Pretty.print ppf "PICTURE@ of@ category@ %a@ is@ incompatible@ with@ \
                         USAGE@ %a; expected@ a@ PICTURE@ for@ %(%)@ data@ item"
         Cobol_data.Picture.pp_category_name ~&picture.category
-        Cobol_ptree.pp_usage_clause usage_clause
+        Cobol_ptree.pp_usage_clause usage
         (match expected with
          | `Numeric_category -> "numeric"
+         | `Numeric_or_alphanum_category -> "numeric@ or@ alphanumeric"
          | `Boolean_class -> "boolean"
          | `Nonalpha_class -> "boolean,@ national,@ national-edited,@ numeric,@ \
                                or@ numeric-edited")
   | Invalid_picture_feature { feature = Digits { given; min; max }
                                       | Scaling { given; min; max } as f;
-                              usage_clause; _ } ->
+                              usage; _ } ->
       Pretty.print ppf "Invalid@ number@ of@ %a@ in@ PICTURE@ for@ item@ \
                         with@ USAGE@ %a; got@ %u,@ expected@ in@ (%u..%u)."
         pp_picture_feature_kind f
-        Cobol_ptree.pp_usage_clause usage_clause
+        Cobol_ptree.pp_usage_clause usage
         given min max
   | Item_not_allowed_in_section { level; section } ->
       Pretty.print ppf "%d-level@ item@ not@ allowed@ in@ %a@ section" ~&level
@@ -290,9 +298,9 @@ let pp_error ppf = function
   | Missing_picture_clause_for_elementary_item { item_name; _ } ->
       Pretty.print ppf "Missing@ PICTURE@ clause@ for@ %a"
         pp_data_name'_opt item_name
-  | Missing_picture_clause_for_item_with_usage { usage_clause; _ } ->
+  | Missing_picture_clause_for_item_with_usage { usage; _ } ->
       Pretty.print ppf "Missing@ PICTURE@ clause@ for@ item@ with@ USAGE@ %a"
-        Cobol_ptree.pp_usage_clause usage_clause
+        Cobol_ptree.pp_usage_clause usage
   | Unexpected_picture_clause { item_name; reason = `Group_item; _ } ->
       Pretty.print ppf "Unexpected@ PICTURE@ clause@ for@ group@ %a"
         pp_data_name'_opt item_name
@@ -337,6 +345,10 @@ let pp_warning ppf = function
   | Ignored_clause_in_redefinition { clause_name; redef_name; _ } ->
       Pretty.print ppf "Ignored@ %s@ clause@ for@ %a@ with@ REDEFINES@ clause"
         clause_name pp_data_name'_opt redef_name
+  | Ignored_picture_clause { usage; _ } ->
+      Pretty.print ppf "Ignored@ PICTURE@ clause@ given@ for@ item@ with@ USAGE@ \
+                        %a"
+        Cobol_ptree.pp_usage_clause usage
   | Mismatching_usage_in_group { item_usage; item_name; group_usage } ->
       Pretty.print ppf "Mismatching@ USAGE@ %a@ for@ %a,@ subordinate@ to@ a@ \
                         group@ with@ USAGE@ %a"
@@ -345,8 +357,8 @@ let pp_warning ppf = function
         Cobol_ptree.pp_usage_clause ~&group_usage
   | Duplicate_clause { clause_name; _ } ->   (* TODO: addendum with second loc *)
       Pretty.print ppf "Duplicate %s clause" clause_name
-  | Unsupported_usage { usage_clause = c } ->
+  | Unsupported_usage { usage } ->
       Pretty.print ppf "Unsupported@ USAGE@ %a"
-        Cobol_ptree.pp_usage_clause ~&c
-  (* | Extraneous_clause { clause_name; _ } -> *)
-  (*     Pretty.print ppf "Extraneous %s clause" clause_name *)
+        Cobol_ptree.pp_usage_clause ~&usage
+(* | Extraneous_clause { clause_name; _ } -> *)
+(*     Pretty.print ppf "Extraneous %s clause" clause_name *)